refactor(contracts): give parse_github_repo_slug one home outside Cortex (#725)
